Moved each layouts into own modules (from jis_support branch)
